la mayoría de la base de datos está hecha solo que no cconsigo acceder a ella y no lo consigo ejecutar correctamente

+++ Supermercado/src/ClasesBase/Producto.java
@@ -65,7 +65,7 @@ public class Producto {
 		return peso;
 	}
 
-	public void setPeso(int peso) {
+	public void setPeso(float peso) {
 		this.peso = peso;
 	}
